requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.*
anyio==4.4.0
attrs==23.2.0
bcrypt==4.1.3
black==24.4.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
cryptography==43.0.0
dataclasses-json==0.6.7
dill==0.3.8
distro==1.9.0
fastapi==0.111.1
frozenlist==1.4.1
graphviz==0.20.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langsmith==0.1.93
llama-index-llms-openai
lxml==5.2.2
Markdown==3.6
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
#nodejs==0.1.1
#npm==0.1.1
numpy>=1.26.0,<2.0.0
omegaconf==2.3.0
openai==1.36.1
optional-django==0.3.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette<=0.38.0
tenacity==8.5.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
word2number==1.1
yarl==1.9.4
